% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NodeAddRemove.R
\name{addNode}
\alias{addNode}
\title{addNode}
\usage{
addNode(input, output, session, ns, network)
}
\arguments{
\item{input}{Shiny input}

\item{output}{Shiny output}

\item{session}{Shiny session}

\item{ns}{The namespace of the visNetwork}

\item{network}{The Network object}
}
\description{
add node interactively
}
