% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function.R
\name{arguments}
\alias{arguments}
\title{arguments}
\usage{
arguments(fn, exclude = NULL)
}
\arguments{
\item{fn}{The function from which the arguments are to be extracted}

\item{exclude}{Arguments to be removed from the extraction}
}
\value{
A list of arguments of the function
}
\description{
Extract arguments from the function.
}
