% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NodeInfoServer.R
\name{edgeDeleteServer}
\alias{edgeDeleteServer}
\title{edgeDeleteServer}
\usage{
edgeDeleteServer(input, network, ns)
}
\arguments{
\item{input}{Shiny input}

\item{network}{A Network object}

\item{ns}{The parent's namespace}
}
\description{
The server side of the edge delete module
}
