% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EdgeModule.R
\name{edgeModule}
\alias{edgeModule}
\title{edgeModule}
\usage{
edgeModule(input, output, session, ns, network)
}
\arguments{
\item{input}{Shiny input}

\item{output}{Shiny output}

\item{session}{Shiny session}

\item{ns}{The namespace of the visNetwork}

\item{network}{The Network object}
}
\description{
edgeModule
}
