% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function.R
\name{extract_input_item}
\alias{extract_input_item}
\title{Extract the items from the function call.}
\usage{
extract_input_item(code_string, fn_args, arg_list)
}
\arguments{
\item{code_string}{The string that contains the function call}

\item{fn_args}{The arguments that are available for this function}

\item{arg_list}{The argument list that the extracted items are to be stored.}
}
\value{
The arg_list
}
\description{
A helper function for the parse_function. Extract individual items
from the function call. The argument names
should be specified for their inputs, although even without the argument names this
function can determine which argument name an input item belongs by relying on the
order and location the input occurs. The first argument is considered the socket unless
a dot is used in another argument. That argument becomes the socket. A name in an input
that is followed by = can either be the argument name or part of the input. If the name is
not in the function call, then it is assumed to be part of the elipses ....
}
