% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkFlow.R
\name{gadgetFlow}
\alias{gadgetFlow}
\title{gadgetFlow}
\usage{
gadgetFlow(.network = NULL)
}
\arguments{
\item{.network}{the network object}
}
\description{
Creates the UI interface of the network using Shiny gadget
}
