% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dictionary.R
\name{is_Dictionary}
\alias{is_Dictionary}
\title{is_Dictionary}
\usage{
is_Dictionary(obj)
}
\arguments{
\item{obj}{The object to be checked}
}
\value{
Logical
}
\description{
Check if the object belongs to the class Dictionary.
}
