% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function.R
\name{is_Function}
\alias{is_Function}
\title{is_Function}
\usage{
is_Function(obj)
}
\arguments{
\item{obj}{The object to be checked}
}
\value{
Logical
}
\description{
Check if the object belongs to the class Function
}
