% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network_Graph.R
\name{is_Network_Graph}
\alias{is_Network_Graph}
\title{is_Network_Graph}
\usage{
is_Network_Graph(obj)
}
\arguments{
\item{obj}{An object}
}
\value{
Logical
}
\description{
Check if the class of the object is Network_Graph
}
