% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkStructure.R
\name{is_Network_Structure}
\alias{is_Network_Structure}
\title{is_Network_Structure}
\usage{
is_Network_Structure(obj)
}
\arguments{
\item{obj}{An object}
}
\value{
Logical
}
\description{
Check if the object is inherited from the Network_Structure
}
