% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkFlow.R
\name{model_flow}
\alias{model_flow}
\title{model_flow}
\usage{
model_flow(.network = NULL)
}
\arguments{
\item{.network}{the network object}
}
\value{
the network object
}
\description{
Opens the UI interface of the network
}
