% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network.R
\name{Network}
\alias{Network}
\title{R6 class representing Network}
\description{
Creates a new network
}
\details{
A network containing nodes and edges
}
\examples{

## ------------------------------------------------
## Method `Network$add_edge`
## ------------------------------------------------

make_data <- function(){ c(3,4,5,NA) }
u<-network_new() \%>\%
      node('data', style = node_style('Dataset'), fn = node_function(fn = 'make_data', lock = T)) \%>\%
      node('mean', style = node_style('Mean'), fn = node_function(fn = 'mean', args = 'x', na.rm = T))
}
\section{Super classes}{
\code{\link[ModelPath:Serializable]{ModelPath::Serializable}} -> \code{\link[ModelPath:Node]{ModelPath::Node}} -> \code{Network}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{structure}}{Return the network structure}

\item{\code{nodes}}{Get or set the nodes}

\item{\code{interrupt}}{Whether the network run has been interrupted.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Network$new()}}
\item \href{#method-reset}{\code{Network$reset()}}
\item \href{#method-clear_outputs}{\code{Network$clear_outputs()}}
\item \href{#method-add_edge}{\code{Network$add_edge()}}
\item \href{#method-remove_edge}{\code{Network$remove_edge()}}
\item \href{#method-lock_all_nodes}{\code{Network$lock_all_nodes()}}
\item \href{#method-add_node}{\code{Network$add_node()}}
\item \href{#method-add_nodes}{\code{Network$add_nodes()}}
\item \href{#method-remove_node}{\code{Network$remove_node()}}
\item \href{#method-build}{\code{Network$build()}}
\item \href{#method-build_graph}{\code{Network$build_graph()}}
\item \href{#method-set_node_validity}{\code{Network$set_node_validity()}}
\item \href{#method-run}{\code{Network$run()}}
\item \href{#method-plot}{\code{Network$plot()}}
\item \href{#method-summary}{\code{Network$summary()}}
\item \href{#method-get_locked_nodes}{\code{Network$get_locked_nodes()}}
\item \href{#method-clone}{\code{Network$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="ModelPath" data-topic="Serializable" data-id="serialize">}\href{../../ModelPath/html/Serializable.html#method-serialize}{\code{ModelPath::Serializable$serialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ModelPath" data-topic="Serializable" data-id="unserialize">}\href{../../ModelPath/html/Serializable.html#method-unserialize}{\code{ModelPath::Serializable$unserialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ModelPath" data-topic="Node" data-id="call">}\href{../../ModelPath/html/Node.html#method-call}{\code{ModelPath::Node$call()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ModelPath" data-topic="Node" data-id="is_locked">}\href{../../ModelPath/html/Node.html#method-is_locked}{\code{ModelPath::Node$is_locked()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ModelPath" data-topic="Node" data-id="lock_to_string">}\href{../../ModelPath/html/Node.html#method-lock_to_string}{\code{ModelPath::Node$lock_to_string()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ModelPath" data-topic="Node" data-id="reset_status">}\href{../../ModelPath/html/Node.html#method-reset_status}{\code{ModelPath::Node$reset_status()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ModelPath" data-topic="Node" data-id="server">}\href{../../ModelPath/html/Node.html#method-server}{\code{ModelPath::Node$server()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ModelPath" data-topic="Node" data-id="set_lock">}\href{../../ModelPath/html/Node.html#method-set_lock}{\code{ModelPath::Node$set_lock()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ModelPath" data-topic="Node" data-id="set_status">}\href{../../ModelPath/html/Node.html#method-set_status}{\code{ModelPath::Node$set_status()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ModelPath" data-topic="Node" data-id="set_ui_editable">}\href{../../ModelPath/html/Node.html#method-set_ui_editable}{\code{ModelPath::Node$set_ui_editable()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ModelPath" data-topic="Node" data-id="status_fail">}\href{../../ModelPath/html/Node.html#method-status_fail}{\code{ModelPath::Node$status_fail()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ModelPath" data-topic="Node" data-id="status_idle">}\href{../../ModelPath/html/Node.html#method-status_idle}{\code{ModelPath::Node$status_idle()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ModelPath" data-topic="Node" data-id="status_running">}\href{../../ModelPath/html/Node.html#method-status_running}{\code{ModelPath::Node$status_running()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ModelPath" data-topic="Node" data-id="status_succeed">}\href{../../ModelPath/html/Node.html#method-status_succeed}{\code{ModelPath::Node$status_succeed()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ModelPath" data-topic="Node" data-id="ui">}\href{../../ModelPath/html/Node.html#method-ui}{\code{ModelPath::Node$ui()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new network
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Network$new(
  id = "",
  style = NULL,
  author = "",
  description = "",
  rankdir = "LR",
  default = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{The id of the Network}

\item{\code{style}}{A Node_Style object}

\item{\code{author}}{Name of the creator of this network}

\item{\code{description}}{Description of this network}

\item{\code{rankdir}}{Direction of the graph; can be 'LR' (left-right) or 'TD' (top-down); default 'LR'}

\item{\code{default}}{Logical. If TRUE, then no data fields are initialized.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A Network object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reset"></a>}}
\if{latex}{\out{\hypertarget{method-reset}{}}}
\subsection{Method \code{reset()}}{
Reset the network by clearing all the outputs of all the
unlocked nodes. Locked nodes are not reset.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Network$reset()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
network
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clear_outputs"></a>}}
\if{latex}{\out{\hypertarget{method-clear_outputs}{}}}
\subsection{Method \code{clear_outputs()}}{
Clears the outputs of all unlocked nodes
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Network$clear_outputs()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_edge"></a>}}
\if{latex}{\out{\hypertarget{method-add_edge}{}}}
\subsection{Method \code{add_edge()}}{
Create an edge between two nodes
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Network$add_edge(representation, socket, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{representation}}{An Edge representation. A character string defined as 'from_node -> to_node'.}

\item{\code{socket}}{Character. The input argument where the 'from' node is plugged into}

\item{\code{...}}{Additional options that need to be pre-specified}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{make_data <- function(){ c(3,4,5,NA) }
u<-network_new() \%>\%
      node('data', style = node_style('Dataset'), fn = node_function(fn = 'make_data', lock = T)) \%>\%
      node('mean', style = node_style('Mean'), fn = node_function(fn = 'mean', args = 'x', na.rm = T))
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-remove_edge"></a>}}
\if{latex}{\out{\hypertarget{method-remove_edge}{}}}
\subsection{Method \code{remove_edge()}}{
Remove an edge
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Network$remove_edge(representation, permanent = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{representation}}{Edge representation. A character string defined as 'from_node -> to_node'.}

\item{\code{permanent}}{Logical; if TRUE, then the edge is removed permanently. If false, the
edge between the two nodes can be rejoined using the function edge_restore}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-lock_all_nodes"></a>}}
\if{latex}{\out{\hypertarget{method-lock_all_nodes}{}}}
\subsection{Method \code{lock_all_nodes()}}{
Lock all the nodes in the network
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Network$lock_all_nodes(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{Logical Set it to TRUE locks all the nodes}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_node"></a>}}
\if{latex}{\out{\hypertarget{method-add_node}{}}}
\subsection{Method \code{add_node()}}{
Add a node to the network
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Network$add_node(node)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{node}}{A Node object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_nodes"></a>}}
\if{latex}{\out{\hypertarget{method-add_nodes}{}}}
\subsection{Method \code{add_nodes()}}{
Add a node to the network
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Network$add_nodes(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Node objects}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-remove_node"></a>}}
\if{latex}{\out{\hypertarget{method-remove_node}{}}}
\subsection{Method \code{remove_node()}}{
Remove a node from the network
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Network$remove_node(node_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{node_id}}{Character. The id of the node to be removed from the network.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-build"></a>}}
\if{latex}{\out{\hypertarget{method-build}{}}}
\subsection{Method \code{build()}}{
Build the configuration of the network for display.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Network$build()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-build_graph"></a>}}
\if{latex}{\out{\hypertarget{method-build_graph}{}}}
\subsection{Method \code{build_graph()}}{
Build the network graph that will be displayed
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Network$build_graph()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_node_validity"></a>}}
\if{latex}{\out{\hypertarget{method-set_node_validity}{}}}
\subsection{Method \code{set_node_validity()}}{
A convenient method that sets the validity of the nodes
in the network. A node is valid if it has a function. Nodes that are valid
will not be locked, whereas nodes that are invalid will be locked. Locking
will also be performed as a side effect.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Network$set_node_validity()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run"></a>}}
\if{latex}{\out{\hypertarget{method-run}{}}}
\subsection{Method \code{run()}}{
Run the network by executing all the node operations sequentially
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Network$run(node_id = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{node_id}}{Character. If the node_id is missing, then all the nodes will be run. Otherwise,
only the node of the node_id will be run.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot"></a>}}
\if{latex}{\out{\hypertarget{method-plot}{}}}
\subsection{Method \code{plot()}}{
Display the network graphically
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Network$plot()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-summary"></a>}}
\if{latex}{\out{\hypertarget{method-summary}{}}}
\subsection{Method \code{summary()}}{
summarize the network
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Network$summary()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
the summary of the network object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_locked_nodes"></a>}}
\if{latex}{\out{\hypertarget{method-get_locked_nodes}{}}}
\subsection{Method \code{get_locked_nodes()}}{
Get all the locked nodes
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Network$get_locked_nodes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A vector of locked node ids
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Network$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
