% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkDisplay.R
\name{networkDisplay}
\alias{networkDisplay}
\title{networkDisplay}
\usage{
networkDisplay(output, ns, network)
}
\arguments{
\item{output}{Shiny output}

\item{ns}{The namespace of the visNetwork}

\item{network}{The network object}
}
\description{
visNetwork display
}
