% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NodeInfoUI.R
\name{nodeDescriptionUI}
\alias{nodeDescriptionUI}
\title{nodeDescriptionUI}
\usage{
nodeDescriptionUI(ns)
}
\arguments{
\item{ns}{The namespace of the visNetwork}
}
\description{
The UI interface of the node description panel
}
