% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NodeInfoServer.R
\name{nodeFunctionServer}
\alias{nodeFunctionServer}
\title{nodeFunctionServer}
\usage{
nodeFunctionServer(input, output, session, node, editable = TRUE)
}
\arguments{
\item{input}{Shiny input}

\item{output}{Shiny output}

\item{session}{Shiny session}

\item{node}{A Node object}

\item{editable}{whether this panel is editable or not}
}
\description{
The server side of the node status function panel module#'
}
