% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NodeInfoUI.R
\name{nodeFunctionUI}
\alias{nodeFunctionUI}
\title{nodeFunctionUI}
\usage{
nodeFunctionUI(ns)
}
\arguments{
\item{ns}{The namespace of the visNetwork}
}
\description{
The UI interface of the node function panel
}
