% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NodeInfoUI.R
\name{nodeInfoUI}
\alias{nodeInfoUI}
\title{nodeInfoUI}
\usage{
nodeInfoUI(ns, ...)
}
\arguments{
\item{ns}{The namespace of the visNetwork}

\item{...}{shiny ui elements}
}
\description{
The UI interface of the nodeInfo module
}
