% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NodeInfoServer.R
\name{nodeOutputTableServer}
\alias{nodeOutputTableServer}
\title{nodeOutputTableServer}
\usage{
nodeOutputTableServer(input, output, session, node)
}
\arguments{
\item{input}{Shiny input}

\item{output}{Shiny output}

\item{session}{Shiny session}

\item{node}{A Node object}
}
\description{
The server side of the node output table panel module
}
