% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NodeInfoUI.R
\name{nodeOutputTableUI}
\alias{nodeOutputTableUI}
\title{nodeIOutputTableUI}
\usage{
nodeOutputTableUI(ns)
}
\arguments{
\item{ns}{The namespace of the visNetwork}
}
\description{
The UI interface of the node output table panel
}
