% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NodeInfoServer.R
\name{nodeStatusServer}
\alias{nodeStatusServer}
\title{nodeStatusServer}
\usage{
nodeStatusServer(input, output, session, node, editable = TRUE, ns)
}
\arguments{
\item{input}{Shiny input}

\item{output}{Shiny output}

\item{session}{Shiny session}

\item{node}{A Node object}

\item{editable}{whether this panel is editable or not}

\item{ns}{The parent's namespace}
}
\description{
The server side of the node status panel module
}
