% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NodeInfoUI.R
\name{nodeVisualizationUI}
\alias{nodeVisualizationUI}
\title{nodeVisualizationUI}
\usage{
nodeVisualizationUI(ns)
}
\arguments{
\item{ns}{The namespace of the visNetwork}
}
\description{
The UI interface of the node visualization panel
}
