% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function.R
\name{parse_expr}
\alias{parse_expr}
\title{Parse a quoted expression}
\usage{
parse_expr(quoted_expr, step_ids, pipe = "\%>\%")
}
\arguments{
\item{quoted_expr}{A quoted expression}

\item{step_ids}{A vector of node ids corresponding to the function calls.}

\item{pipe}{The pipe that separates individual function calls. Default is %>%.}
}
\value{
A list of parsed function items.
}
\description{
This function parsed an expression into individual function calls. Each
function call is further parsed to obtain the namespace of the function, the function,
the arguments and their inputs. Individual function calls must be connected by the pipe such as
the default %>%. The whole expression must be quoted i.e. quote(expr). The step_ids
must be specified for each function call and these ids will become their node ids.
}
