% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileIO.R
\name{readOutput}
\alias{readOutput}
\title{Read output data from a file.}
\usage{
readOutput(filename = NULL, repo = c("local"))
}
\arguments{
\item{filename}{The name of the file}

\item{repo}{The type of repository from which the data will be loaded}
}
\value{
The output data
}
\description{
Load the output data from a file
}
