% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileIO.R
\name{readR6}
\alias{readR6}
\title{Load a file containing the R6 object}
\usage{
readR6(filename = "", obj)
}
\arguments{
\item{filename}{The name of the file to read}
}
\value{
The saved R6 object
}
\description{
Load an R6 object from a RDS file.
}
