% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileIO.R
\name{saveOutput}
\alias{saveOutput}
\title{Save the output data into a file.}
\usage{
saveOutput(x, filename = NULL, repo = c("local"), type = c("RDS", "csv"))
}
\arguments{
\item{x}{The output data to be saved}

\item{filename}{The name of the file}

\item{repo}{The type of repository where the data will be saved.}

\item{type}{The type of file to be saved}
}
\description{
Save the output created by a node into a file.
}
