% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Serializable.R
\name{Serializable}
\alias{Serializable}
\title{R6 Class representing a Serializable class}
\description{
A Serializable is the parent of a R6 inheritance tree.
}
\details{
A Serializable contains methods that serialize the objects and restore the byte codes back into R objects.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-serialize}{\code{Serializable$serialize()}}
\item \href{#method-unserialize}{\code{Serializable$unserialize()}}
\item \href{#method-clone}{\code{Serializable$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-serialize"></a>}}
\if{latex}{\out{\hypertarget{method-serialize}{}}}
\subsection{Method \code{serialize()}}{
Serialize this object. Only the data fields are serialized.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Serializable$serialize()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A raw object.
unserialize
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-unserialize"></a>}}
\if{latex}{\out{\hypertarget{method-unserialize}{}}}
\subsection{Method \code{unserialize()}}{
Unserialize the raw data and restore the data fields.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Serializable$unserialize(s)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{s}}{A raw serialized object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Serializable$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
