% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyFlow.R
\name{shinyFlowServer}
\alias{shinyFlowServer}
\title{shinyFlowServer}
\usage{
shinyFlowServer(
  input,
  output,
  session,
  network,
  network_options = list(lockable = TRUE, edgeAddDelete = TRUE, fileMenu = TRUE,
    nodeAddDelete = TRUE, description = TRUE)
)
}
\arguments{
\item{input}{the shiny input}

\item{output}{the shiny output}

\item{session}{the shiny session}

\item{series}{the series containing the network and the graph}

\item{network_Options}{the UI features that should be hidden from the user. These features are:
nodeAddDelete Add and delete a node in the interface
edgeAddDelete Add and delete an edge
fileLoadSave Load and save network
description Change network description
lockable Ability to lock individual locks}
}
\description{
The call module of shinyFlowUI. Use callModule to call this function
}
