% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkDisplay.R
\name{updateNetworkStatus}
\alias{updateNetworkStatus}
\title{networkDisplay}
\usage{
updateNetworkStatus(graph, ns)
}
\arguments{
\item{graph}{The Network_Graph object}

\item{ns}{The namespace of the visNetwork}
}
\description{
visNetwork display
}
