% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkDisplay.R
\name{updateNodeStatus}
\alias{updateNodeStatus}
\title{updateNodeStatus}
\usage{
updateNodeStatus(node_table, ns)
}
\arguments{
\item{node_table}{The node table for rendering the visNetwork}

\item{ns}{The namespace of the visNetwork}
}
\description{
Update visNetwork node display
}
