% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkDisplay.R
\name{visNetworkShow}
\alias{visNetworkShow}
\title{visNetworkShow}
\usage{
visNetworkShow(graph, dir = "LR")
}
\arguments{
\item{graph}{The Network_Graph object}

\item{dir}{The direction of the network}
}
\description{
Show the visNetwork graph
}
