% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileIO.R
\name{writeR6}
\alias{writeR6}
\title{Write R6 object into file}
\usage{
writeR6(obj, filename = "")
}
\arguments{
\item{obj}{The object to serialize.}

\item{filename}{The name of the file}
}
\description{
Write an R6 object into a RDS file.
}
